requirements.txt
absl-py
tensorflow==2.3.0rc0
Pillow
opencv-python==4.1.1.26
numpy==1.16.0
Flask
Werkzeug
gunicorn==19.9.0
easydict==1.9
